test for some excuse find testing
